move the save logic to the close button
