+++ src/components/Map/CustomDataInventory/HazardLayerList.tsx
@@ -11,11 +11,6 @@ import { Typography } from "@mui/joy";
 import { LayerItem } from "@app/components/Map/CustomDataInventory/LayerItem";
 import config from "@app/app.config";
 
-// type HazardLayerListProps = {
-//     visualization: Visualization;
-// };
-
-// export const HazardLayerList = ({ visualization }: HazardLayerListProps) => {
 export const HazardLayerList = () => {
     const hazards = useSelector((state: RootState) => state.project.projectHazards);
 
@@ -97,7 +92,6 @@ export const HazardLayerList = () => {
 
                                             return dataset ? (
                                                 <Box key={dataset.id} className="flex justify-center px-[32px]">
-                                                    {/* <LayerItem dataset={dataset} visualization={visualization} /> */}
                                                     <LayerItem dataset={dataset} />
                                                 </Box>
                                             ) : null;
