+++ src/components/Map/CustomDataInventory/LayerItem.tsx
@@ -4,14 +4,9 @@ import { IconButton } from "@mui/material";
 import { AddCircleOutline, RemoveCircle, DescriptionOutlined } from "@mui/icons-material";
 
 import { useSelector, useDispatch, addLayer, removeLayer, selectDataset, RootState } from "@ncsa/geo-explorer";
-// import { getHeaders, inferLayerType } from "@app/utils";
-// import { useParams } from "react-router-dom";
-// import axios from "axios";
-// import config from "@app/app.config";
 
 export type DatasetLayerItemProps = {
     dataset: Dataset;
-    // visualization: Visualization;
 };
 
 export const LayerItem = ({ dataset }: DatasetLayerItemProps) => {
@@ -23,49 +18,10 @@ export const LayerItem = ({ dataset }: DatasetLayerItemProps) => {
     const handleSelect = () => geoExplorerDispatch(selectDataset({ layer_id: dataset.id }));
 
     const handleAdd = async () => {
-        // const layers = [
-        //     {
-        //         workspace: "incore",
-        //         layerId: dataset.id,
-        //         displayName: dataset.title,
-        //         description: dataset.description,
-        //         datasetCategoryType: dataset.dataType,
-        //         layerType: inferLayerType(dataset.dataType),
-        //         boundingBox: dataset.boundingBox
-        //     }
-        // ];
-        //
-        // if (!id) {
-        //     console.error("Project ID is not defined. Cannot add layer to visualization.");
-        //     return;
-        // }
-        //
-        // try {
-        //     await axios.post(`${config.projectApi}/${id}/visualizations/${visualization.id}/layers`, layers, {
-        //         headers: getHeaders()
-        //     });
-        //     console.log("Layer added to visualization successfully.");
-        // } catch (error) {
-        //     console.error("Failed to add layer to visualization:", error);
-        // }
         geoExplorerDispatch(addLayer({ layer_id: dataset.id }));
     };
 
     const handleRemove = async () => {
-        // if (!id) {
-        //     console.error("Project ID is not defined. Cannot delete layer from visualization.");
-        //     return;
-        // }
-        //
-        // try {
-        //     await axios.delete(`${config.projectApi}/${id}/visualizations/${visualization.id}/layers`, {
-        //         headers: getHeaders(),
-        //         data: [dataset.id]
-        //     });
-        //     console.log("Layer deleted from visualization successfully.");
-        // } catch (error) {
-        //     console.error("Failed to delete layer from visualization:", error);
-        // }
         geoExplorerDispatch(removeLayer({ layer_id: dataset.id }));
     };
