+++ src/reducer/projectSlice.ts
@@ -28,12 +28,12 @@ const initialState: ProjectState = {
 export const getProjects = createAsyncThunk(
     "projects/getProjects",
     async ({
-               skip = 0,
-               limit = 9,
-               filters = {},
-               sortBy = "date",
-               order = "desc"
-           }: {
+        skip = 0,
+        limit = 9,
+        filters = {},
+        sortBy = "date",
+        order = "desc"
+    }: {
         skip?: number;
         limit?: number;
         filters?: Record<string, string | number>;
@@ -66,9 +66,7 @@ export const getProjects = createAsyncThunk(
     }
 );
 
-export const createProject = createAsyncThunk("projects/createProject", async ({ project }: {
-    project: ProjectIn
-}) => {
+export const createProject = createAsyncThunk("projects/createProject", async ({ project }: { project: ProjectIn }) => {
     const response = await axios.post(`${PROJECT_API_URL}`, project, {
         headers: getHeaders()
     });
@@ -103,11 +101,11 @@ export const deleteProject = createAsyncThunk("projects/deleteProject", async (p
 export const editProject = createAsyncThunk(
     "projects/editProject",
     async ({
-               projectId,
-               name,
-               description,
-               region
-           }: {
+        projectId,
+        name,
+        description,
+        region
+    }: {
         projectId: string;
         name: string;
         description: string;
@@ -132,13 +130,13 @@ export const editProject = createAsyncThunk(
 export const getProjectDatasets = createAsyncThunk(
     "projects/getProjectDatasets",
     async ({
-               projectId,
-               skip = 0,
-               limit = 9,
-               filters = {},
-               sortBy = "date",
-               order = "desc"
-           }: {
+        projectId,
+        skip = 0,
+        limit = 9,
+        filters = {},
+        sortBy = "date",
+        order = "desc"
+    }: {
         projectId: string;
         skip?: number;
         limit?: number;
@@ -174,13 +172,13 @@ export const getProjectDatasets = createAsyncThunk(
 export const getProjectHazards = createAsyncThunk(
     "projects/getProjectHazards",
     async ({
-               projectId,
-               skip = 0,
-               limit = 9,
-               filters = {},
-               sortBy = "date",
-               order = "desc"
-           }: {
+        projectId,
+        skip = 0,
+        limit = 9,
+        filters = {},
+        sortBy = "date",
+        order = "desc"
+    }: {
         projectId: string;
         skip?: number;
         limit?: number;
@@ -216,13 +214,13 @@ export const getProjectHazards = createAsyncThunk(
 export const getProjectDRF3Mappings = createAsyncThunk(
     "projects/getProjectDRF3Mappings",
     async ({
-               projectId,
-               skip = 0,
-               limit = 9,
-               filters = {},
-               sortBy = "date",
-               order = "desc"
-           }: {
+        projectId,
+        skip = 0,
+        limit = 9,
+        filters = {},
+        sortBy = "date",
+        order = "desc"
+    }: {
         projectId: string;
         skip?: number;
         limit?: number;
@@ -258,13 +256,13 @@ export const getProjectDRF3Mappings = createAsyncThunk(
 export const getProjectWorkflows = createAsyncThunk(
     "projects/getProjectWorkflows",
     async ({
-               projectId,
-               skip = 0,
-               limit = 9,
-               filters = {},
-               sortBy = "date",
-               order = "desc"
-           }: {
+        projectId,
+        skip = 0,
+        limit = 9,
+        filters = {},
+        sortBy = "date",
+        order = "desc"
+    }: {
         projectId: string;
         skip?: number;
         limit?: number;
@@ -300,13 +298,13 @@ export const getProjectWorkflows = createAsyncThunk(
 export const getProjectVisualizations = createAsyncThunk(
     "projects/getProjectVisualizations",
     async ({
-               projectId,
-               skip = 0,
-               limit = 9,
-               filters = {},
-               sortBy = "date",
-               order = "desc"
-           }: {
+        projectId,
+        skip = 0,
+        limit = 9,
+        filters = {},
+        sortBy = "date",
+        order = "desc"
+    }: {
         projectId: string;
         skip?: number;
         limit?: number;
@@ -340,9 +338,9 @@ export const getProjectVisualizations = createAsyncThunk(
 );
 
 export const fetchInfiniteProjectVisualizations = async ({
-                                                             pageParam = 1,
-                                                             projectId
-                                                         }: any): Promise<{
+    pageParam = 1,
+    projectId
+}: any): Promise<{
     data: Visualization[];
     nextPage: number;
 }> => {
@@ -477,10 +475,10 @@ export const createProjectVisualization = createAsyncThunk(
 export const addLayerToVisualization = createAsyncThunk(
     "projects/addLayerToVisualization",
     async ({
-               projectId,
-               visualizationId,
-               layers
-           }: {
+        projectId,
+        visualizationId,
+        layers
+    }: {
         projectId: string;
         visualizationId: string;
         layers: IncoreLayer[];
@@ -499,10 +497,10 @@ export const addLayerToVisualization = createAsyncThunk(
 export const deleteLayerToVisualization = createAsyncThunk(
     "projects/deleteLayerToVisualization",
     async ({
-               projectId,
-               visualizationId,
-               layerIds
-           }: {
+        projectId,
+        visualizationId,
+        layerIds
+    }: {
         projectId: string;
         visualizationId: string;
         layerIds: string[];
@@ -521,10 +519,10 @@ export const deleteLayerToVisualization = createAsyncThunk(
 export const patchVisualization = createAsyncThunk(
     "projects/patchVisualization",
     async ({
-               projectId,
-               visualizationId,
-               patchData
-           }: {
+        projectId,
+        visualizationId,
+        patchData
+    }: {
         projectId: string;
         visualizationId: string;
         patchData: {
